requirements.txt
Flask==2.3.0
groq==0.9.0
python-dotenv==1.0.0
Werkzeug==2.3.0
gunicorn==21.2.0
